requirements.txt
optuna
transformers==4.46.3
torch==2.4.1
torchvision==0.19.1
matplotlib==3.7.5
mlflow==2.17.2
timm==1.0.15
sentence-transformers==4.1.0
python-dotenv==1.1.0
torchviz==0.0.3
